update of ERB (Electron & other deps with tooling)

+++ .erb/scripts/clean.js
@@ -1,4 +1,7 @@
+import fs from 'fs';
+
 import rimraf from 'rimraf';
+
 import webpackPaths from '../configs/webpack.paths';
 
 const foldersToRemove = [
@@ -8,5 +11,5 @@ const foldersToRemove = [
 ];
 
 foldersToRemove.forEach((folder) => {
-  rimraf.sync(folder);
+  if (fs.existsSync(folder)) rimraf.sync(folder);
 });
